requirements.txt
Pillow
astropy
numpy
matplotlib
